Move _get_hypervolume_history_info, _HypervolumeHistoryInfo, and tests for hypervolume history plot from optuna.visualization.matplotlib to optuna.visualization.
